Added data module and code to get all estimators
