+++ .gitignore
@@ -1,4 +1,5 @@
 data/
+!src/data/
 spectral-metric/
 .ruff_cache/
 __pycache__/
